Renamed packages containing "module"
