+++ src/guis/zeale/guis/developer_page/ConsoleModule.java
@@ -1,4 +1,4 @@
-package zeale.guis.developer_module;
+package zeale.guis.developer_page;
 
 import java.io.IOException;
 import java.io.OutputStream;
